+++ src/io/trr_io.rs
@@ -458,7 +458,7 @@ impl PrivateTrajWrite for TrrWriter {
             Some(x) => system
                 .get_groups()
                 .get(x)
-                .ok_or_else(|| WriteTrajError::GroupNotFound(x.to_owned()))?
+                .map_err(|_| WriteTrajError::GroupNotFound(x.to_string()))?
                 .clone(),
             None => system
                 .get_groups()