+++ src/io/xtc_io/mod.rs
@@ -277,7 +277,7 @@ mod xtc_write {
                 Some(x) => system
                     .get_groups()
                     .get(x)
-                    .ok_or_else(|| WriteTrajError::GroupNotFound(x.to_owned()))?
+                    .map_err(|_| WriteTrajError::GroupNotFound(x.to_owned()))?
                     .clone(),
                 None => system
                     .get_groups()